scripts/requirements.txt
psycopg2-binary==2.9.12
requests==2.32.3
beautifulsoup4==4.14.3
